update entry list items from Firestore when editor
