+++ src/lib/stores/index.ts
@@ -1,5 +1,5 @@
 import { admin } from './admin';
-import { isManager, isContributor } from './editing';
+import { isManager, isContributor, canEdit } from './editing';
 
 import { myDictionaries } from './dictionaries';
 import { dictionary, entries } from './dictionary';
@@ -10,6 +10,7 @@ export {
   admin,
   isManager,
   isContributor,
+  canEdit,
   myDictionaries,
   dictionary,
   entries,
